+++ Agemarker/Events/ResultsFileLoadedEventArgs.cs
@@ -10,11 +10,11 @@ namespace Agemarker.Events
         public double[] OxidesContent { get; set; }
         public double[] ElementsContent { get; set; }
         public double[] ElementsWeight { get; set; }
-        public int Multiplier { get; set; }
+        public long Multiplier { get; set; }
         public int IntervalsNumber { get; set; }
         public AgemarkerCore.Data.Logarithm Log { get; set; }
 
-        public ResultsFileLoadedEventArgs(double[] oxidesContent, double[] elementsContent, double[] elementsWeight, int multiplier, int intervalsNumber, AgemarkerCore.Data.Logarithm log)
+        public ResultsFileLoadedEventArgs(double[] oxidesContent, double[] elementsContent, double[] elementsWeight, long multiplier, int intervalsNumber, AgemarkerCore.Data.Logarithm log)
         {
             this.OxidesContent = oxidesContent;
             this.ElementsContent = elementsContent;
